feat(frontend): enhance rss feed input and component
