+++ apps/frontend/app/app/(monitor)/rss-feed-config/styles.ts
@@ -19,6 +19,21 @@ export default StyleSheet.create({
     padding: 10,
     fontFamily: 'Poppins_400Regular',
   },
+  example: {
+    marginTop: 4,
+    fontSize: 12,
+    fontFamily: 'Poppins_400Regular',
+  },
+  addButton: {
+    padding: 10,
+    borderRadius: 8,
+    alignItems: 'center',
+    marginBottom: 20,
+  },
+  addButtonText: {
+    fontSize: 14,
+    fontFamily: 'Poppins_600SemiBold',
+  },
   button: {
     padding: 15,
     borderRadius: 8,
